tests/integration/requirements.txt
playwright==1.41.2
pytest==8.0.0
pytest-playwright==0.4.4
python-dotenv==1.0.1
